# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2.2.0
  description: Construction AI Platform — Phase 2 API (routes match FastAPI backend exactly)
servers:
  - url: /api
    description: Base API path (Python FastAPI backend)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      required: [id, doc_type, title, doc_date, content_summary, is_archived]
      properties:
        id:
          type: integer
        project_id:
          type: ["integer", "null"]
        organization_id:
          type: ["integer", "null"]
        doc_type:
          type: string
        title:
          type: string
        doc_date:
          type: string
        content_summary:
          type: string
        original_filename:
          type: ["string", "null"]
        mime_type:
          type: ["string", "null"]
        file_size:
          type: ["integer", "null"]
        checksum:
          type: ["string", "null"]
        uploaded_at:
          type: ["string", "null"]
          format: date-time
        uploaded_by:
          type: ["integer", "null"]
        updated_at:
          type: ["string", "null"]
          format: date-time
        version_number:
          type: ["integer", "null"]
        is_archived:
          type: boolean
    DocumentVersion:
      type: object
      required: [version_number, original_filename, mime_type, file_size, checksum, uploaded_at, is_current]
      properties:
        version_number:
          type: integer
        original_filename:
          type: string
        mime_type:
          type: string
        file_size:
          type: integer
        checksum:
          type: string
        uploaded_at:
          type: string
          format: date-time
        uploaded_by:
          type: ["integer", "null"]
        is_current:
          type: boolean
    DocumentVersionUpload:
      allOf:
        - $ref: "#/components/schemas/DocumentVersion"
        - type: object
          required: [is_duplicate]
          properties:
            is_duplicate:
              type: boolean

    # ── Admin — user directory (existing, admin-role-only endpoint) ─────
...
security:
  - bearerAuth: []
paths:
...
  /v1/documents/{document_id}/versions:
    get:
      operationId: listDocumentVersions
      tags: [documents]
      summary: List a document's version history, newest first
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"
        "404":
          description: Not found
    post:
      operationId: uploadDocumentVersion
      tags: [documents]
      summary: Upload a new version of a document's file
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: New (or, if byte-identical, existing) current version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersionUpload"
        "404":
          description: Not found
  /v1/documents/{document_id}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Download the current (or a specific ?version=N) file
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
        - name: version
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
  /v1/documents/{document_id}/archive:
    post:
      operationId: archiveDocument
      tags: [documents]
      summary: Soft-archive a document (nothing is ever permanently removed)
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
  /v1/documents/{document_id}/unarchive:
    post:
      operationId: unarchiveDocument
      tags: [documents]
      summary: Unarchive a document
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found

  # ── Claims — read-only source-entity context for approvals ───────────
...